// src/models/ProductSearchQuery.ts
/*
 * Copyright (c) 2022 The Trebol eCommerce Project
 *
 * This software is released under the MIT License.
 * https://opensource.org/licenses/MIT
 */

export class ProductSearchQuery {
  nameLike?: string;
  categoryCode?: string;
  // tags?: string;
  // brand?: string;
  // brandLike?: string;
}